+++ packages/core/babel-preset-dlight/src/nodeHelper.ts
@@ -151,3 +151,13 @@ export function isDLightView(path: any) {
 
   return t.isIdentifier(node.superClass, { name: "View" })
 }
+
+export function arrowFunctionPropertyToMethod(propertyNode: t.ClassProperty) {
+  const propertyBody = (propertyNode.value as t.ArrowFunctionExpression).body
+  const body = t.isExpression(propertyBody) ? t.blockStatement([t.returnStatement(propertyBody)]) : propertyBody
+  propertyNode.type = "ClassMethod" as any
+  (propertyNode as any).body = body
+  ;(propertyNode as any).kind = "method"
+  ;(propertyNode as any).params = (propertyNode.value as t.ArrowFunctionExpression).params
+  propertyNode.value = null
+}