+++ packages/core/babel-preset-dlight/CHANGELOG.md
@@ -1,5 +1,11 @@
 # babel-preset-dlight
 
+## 0.9.29
+
+### Patch Changes
+
+- feat: add subview type support
+
 ## 0.9.28
 
 ### Patch Changes