+++ packages/core/types/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @dlightjs/types
 
+## 0.9.30
+
+### Patch Changes
+
+- Updated dependencies
+  - @dlightjs/dlight@0.9.30
+
 ## 0.9.29
 
 ### Patch Changes